refactor: move modals
